requirements.txt
aiohttp
aiohttp-jinja2
aiohttp-jinja2>=1.5
aiohttp>=3.8.0
certifi>=2023.7.22
deep-translator>=1.11.0
discord.py>=2.3.0
feedparser>=6.0.0
jinja2>=3.1.2
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
python-dateutil>=2.8.2
requests>=2.31.0